refactor: improve registry implementation
Summary of changes:
* removed Registry interface as there is only a single implementation
* the registry now receives a stub instance at instantiation and there
  is no Context parameter to pass for each of its methods
* the registry subpackage has been removed

Rationale for the second point:
There is no use case for a Registry instance that has a lifecycle in
which it will be used with multiple Context (or ChaincodeStub)
instances.  The Registry's lifecycle follows that of the Context.
Still, Registry could be a utility class with static methods receiving
a Context or ChaincodeStub parameter, but this would be leaning towards
FP and we are in Java. The OO way is what this commit introduces.

Signed-off-by: bzp99 <bpeter@edu.bme.hu>