fix: create registry dynamically in context
We cannot create a Registry instance at construction time of
ContextWithRegistry, because Registry needs a stub instance.  However,
the 'head' stub instance is unknown at the construction of
ContextWithRegistry (normally, from a child class).

At the same time, it makes sense to create the Registry on the fly.
Context is an ephemeral object anyway and normally, one would only
invoke getRegistry once and then reuse that.